TINYMCE-14504: rename main to Main
