+++ modules/oxide-components/src/main/ts/components/iconbutton/IconButton.stories.tsx
@@ -1,6 +1,7 @@
 import { Obj } from '@ephox/katamari';
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { UniverseProvider } from 'oxide-components/main';
+
+import { UniverseProvider } from '../../contexts/universecontext/UniverseProvider';
 
 import { IconButton } from './IconButton';
